Remove maxRequests circuit breaker type
It seems like maxRequests isn't an ideal place for the circuit breaker.
In case user really wants to circuit break because upstream service
won't be able to handle large no. of requests, then it should be
configured based on other circuit breaker.

For overall max no. of requests to the upstream cluster, it should be
mentioned somewhere else, but not in circuit breaker configuration.

Signed-off-by: Ashish Bansal <bansal.ashish096@gmail.com>